+++ src/hooks/FLAlertLayer.hpp
@@ -1,11 +1,56 @@
 #pragma once
 
+#include <Geode/modify/FLAlertLayer.hpp>
+
 #include <Geode/ui/Button.hpp>
 #include <Geode/ui/TextArea.hpp>
-#include "Global.hpp"
-#include "DeltaruneTextArea.hpp"
-#include "ImageNode.hpp"
+#include "../DeltaruneTextArea.hpp"
+#include "../ImageNode.hpp"
 #include <alphalaneous.alphas-ui-pack/include/API.hpp>
+#include "../Events.hpp"
+
+namespace DeltaruneMaps {
+	enum Character {
+		DEFAULT,
+		TYPEWRITER,
+		TORIEL,
+		SANS,
+		PAPYRUS,
+		UNDYNE,
+		ALPHYS,
+		ASGORE,
+		ASRIEL,
+		SUSIE,
+		RALSEI,
+		LANCER,
+		JEVIL,
+		NOELLE,
+		BERDLY,
+		QUEEN,
+		SPAMTON,
+		SPAMTON_NEO,
+		TENNA,
+		CAROL,
+		GERSON,
+		JACKENSTEIN,
+		FLOWERY,
+		NUM_CHARACTERS,
+		NONE // This is after NUM_CHARACTERS intentionally because it's a special case
+	};
+	
+	struct CharacterData {
+		std::string sound = "SND_TXT1";
+		int soundRate = 2;
+		bool hasPitchVariation = false;
+	};
+
+	inline std::unordered_map<Character, CharacterData> characterToData;
+	inline std::unordered_map<std::string, Character> nameToCharacter;
+	inline std::unordered_map<std::string, Character> titleToCharacter;
+
+	void init();
+}
+
 class $modify(DeltaruneAlertLayer, FLAlertLayer) {
 	struct Fields {
 		bool clickedChoice = false;
@@ -18,7 +63,7 @@ class $modify(DeltaruneAlertLayer, FLAlertLayer) {
 		bool doneRolling = false;
 		bool rolledPage = false;
 		bool done = false;
-		bool disableClickToProgress = Mod::get()->getSettingValue<bool>("disableClickToProgress");
+		bool disableClickToProgress = geode::Mod::get()->getSettingValue<bool>("disableClickToProgress");
 		float screenSize = 0.f;
 		float lostTime = 0;
 		float contentXOffset = 0.f;
@@ -35,14 +80,14 @@ class $modify(DeltaruneAlertLayer, FLAlertLayer) {
 		int soundTimer = 0;
 		CCMenuItemSpriteExtra* old_btn1 = nullptr;
 		CCMenuItemSpriteExtra* old_btn2 = nullptr;
-		Button* btn1 = nullptr;
-		Button* btn2 = nullptr;
+		geode::Button* btn1 = nullptr;
+		geode::Button* btn2 = nullptr;
 		CCNode* textContentNode = nullptr;
 		TextArea* old_textArea = nullptr;
 		std::shared_ptr<DeltaruneTextArea> m_textArea = nullptr;
-		CCScale9Sprite* bg = nullptr;
-		CCLabelBMFont* title = nullptr;
-		CCSprite* heart = nullptr;
+		cocos2d::extension::CCScale9Sprite* bg = nullptr;
+		cocos2d::CCLabelBMFont* title = nullptr;
+		cocos2d::CCSprite* heart = nullptr;
 		alpha::ui::RenderNode* renderedSprite = nullptr;
 		ImageNode* imageNode;
 		DialogLayer* dialogLayer;
@@ -52,10 +97,8 @@ class $modify(DeltaruneAlertLayer, FLAlertLayer) {
 		std::vector<std::string> characterSpriteNames;
 		std::vector<std::string> titles;
 		std::string text = "";
-		std::string textSound = Mod::get()->getSettingValue<std::string>("textSound");
-		std::unordered_map<std::string, std::string> nameToFile;
-		std::unordered_map<std::string, std::string_view> nameToSound;
-		std::unordered_map<std::string, int> nameToSoundRate;
+		DeltaruneMaps::Character textSound = DeltaruneMaps::Character::NONE;
+		std::vector<DeltaruneDialogObject> nextAlerts;
 	};
 	void registerKeybinds();
 	void animateBG(float);
@@ -64,17 +107,16 @@ class $modify(DeltaruneAlertLayer, FLAlertLayer) {
 	void changeButtons();
 	void changeTitle();
 	void handleSound();
-	CCLabelBMFont* createStar();
+	cocos2d::CCLabelBMFont* createStar();
 	void changeText();
 	void changeLook();
 	void decideToBlockKeys();
 	void show() override;
 	void initSoundRate();
-	void initMaps();
 	bool init(FLAlertLayerProtocol*, char const*, gd::string, char const*, char const*, float, bool, float, float);
 	void rollText(float);
 	void showButtons();
-	void setHeartPosition(Button*);
+	void setHeartPosition(geode::Button*);
 	int getLinesLeft();
 	void skipText();
 	int emptyLinesAmount(int = 0);
@@ -84,9 +126,11 @@ class $modify(DeltaruneAlertLayer, FLAlertLayer) {
 	void pickChoice();
 	void onBtn2(CCObject*);
 	void onBtn1(CCObject*);
-	void clickedOnButton(Button*, Button*, int);
-	bool ccTouchBegan(CCTouch*, CCEvent*) override;
+	void clickedOnButton(geode::Button*, geode::Button*, int);
+	bool ccTouchBegan(cocos2d::CCTouch*, cocos2d::CCEvent*) override;
 	void playSound(char);
 	void updateRenderTexture();
-	void setTextSound(std::string const& textSound);
+	void setTextSound(DeltaruneMaps::Character textSound);
+	// A function to execute after the textbox is closed
+	void setNextAlerts(std::vector<DeltaruneDialogObject> const& nextAlerts);
 };